bot hooks
